practical-penguins/requirements.txt
django==3.0.8
wheel
Pillow==7.2.0
django-crispy-forms
django-formset-js
twilio==6.44.2
django-eventstream
django-phonenumber-field
phonenumbers
django-dotenv